+++ Tests/config/common/config_behat.php
@@ -1,8 +1,15 @@
 <?php
 
-// Used to set parameters based on env vars, regardless of the SF version in use
+// Used to set parameters based on env vars, regardless of the SF and eZ versions in use
 
-$container->setParameter('database_host', getenv('MYSQL_HOST'));
-$container->setParameter('database_user', getenv('MYSQL_USER'));
-$container->setParameter('database_password', getenv('MYSQL_PASSWORD'));
-$container->setParameter('database_name', getenv('MYSQL_DATABASE'));
+// we rely on Doctrine aliases for this to work
+$container->setParameter('database_driver', getenv('DB_TYPE'));
+// and on Docker conatienrs configuration for having hostnames that match db types
+if (($dbHost = getenv('DB_HOST')) !== false) {
+    $container->setParameter('database_host', $dbHost);
+} else {
+    $container->setParameter('database_host', getenv('DB_TYPE'));
+}
+$container->setParameter('database_user', getenv('DB_EZ_USER'));
+$container->setParameter('database_password', getenv('DB_EZ_PASSWORD'));
+$container->setParameter('database_name', getenv('DB_EZ_DATABASE'));